chore: align compiler test
