+++ packages/rspack/jest.config.js
@@ -5,13 +5,13 @@ const config = {
 	testEnvironment: "../../scripts/test/patch-node-env.cjs",
 	setupFilesAfterEnv: ["<rootDir>/tests/setupTestFramework.js"],
 	testMatch: [
-		"<rootDir>/tests/Compiler.test.js"
-		// "<rootDir>/tests/*.basictest.ts",
-		// "<rootDir>/tests/*.basictest.js",
-		// "<rootDir>/tests/*.longtest.ts",
-		// "<rootDir>/tests/*.unittest.ts",
-		// "<rootDir>/tests/copyPlugin/*.test.js",
-		// "<rootDir>/tests/WatchSuspend.test.js"
+		"<rootDir>/tests/*.test.ts",
+		"<rootDir>/tests/*.basictest.ts",
+		"<rootDir>/tests/*.basictest.js",
+		"<rootDir>/tests/*.longtest.ts",
+		"<rootDir>/tests/*.unittest.ts",
+		"<rootDir>/tests/copyPlugin/*.test.js",
+		"<rootDir>/tests/WatchSuspend.test.js"
 	],
 	testTimeout: process.env.CI ? 60000 : 30000,
 	cache: false,
